VK-720 Use service-versions to trigger activating-flexibility-tests
As part of VK-720 we added an extra parameter to the activating-flexibility-tests to be able to receive more versions (we hit the 10 inputs limit on GitHub.

We are now using the new map in place of the old service+version: this means that the repositories that are using the old version of this composite action can continue to build with what they have, but we are encouraging the migration to the service-versions map when updating this reference.