VK-720 Use service-versions to trigger activating-flexibility-tests
As part of VK-720 we added an extra parameter to the activating-flexibility-tests to be able to receive more versions (we hit the 10 inputs limit on GitHub.

Had a chat with Skar, and we are keeping the interface of one "service / one version" towards the client and using the map internally.

This requires a change also in the activating-flexibility-tests to deal with the map for every service